+++ src/main/java/cat/cultura/backend/repository/EventJpaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface EventJpaRepository extends JpaRepository<Event, Long> {
-    Event findByCodi(int codi);
+    Event findByCodi(Long codi);
 }
